store da-ghc-lib on a GCS bucket
At the moment, publishing a new version of da-ghc-lib (the one DAML
depends on) is a manual process aimed at Bintray. As we have just
received a note from JFrog that Bintray is sunsetting, this is a good
opportunity to automate the deployment.

This PR aims at creating a GCS bucket (Terraform files to be applied
manually) and change the CI process to push every master commit to said
GCS bucket.