more flexible da flavour
This PR updates the `CI.hs` script to make the DAML-specific build more
flexible. The motivation for this change is to set up a fully auditable
CI process for the DAML repo, i.e. one in which we can specify (and
track) exactly which commits were involved in creating the version of
ghc-lib used by the DAML compiler.